+++ app/directive-test.component.ts
@@ -1,9 +1,9 @@
-import { Http } from "@angular/http";
-import { Observable } from 'rxjs/Observable';
-import { Component, ViewEncapsulation } from "@angular/core";
-import { DomSanitizer, SafeHtml } from "@angular/platform-browser";
+import {Http} from "@angular/http";
+import {Component, ViewEncapsulation} from "@angular/core";
+import {DomSanitizer, SafeHtml} from "@angular/platform-browser";
 
-import { AppSvc } from "./app.service";
+import {AppSvc} from "./app.service";
+
 let templateStr: string = `
   <h1> Autocomplete Directive Test - Local Source </h1>
     
@@ -11,8 +11,11 @@ let templateStr: string = `
     <ngui-utils-1>
       <div ngui-auto-complete 
         [source]="arrayOfStrings"
-        [accept-user-input]="false"
+        [accept-user-input]="true"
+        [auto-select-first-item]="false"
+        [select-on-blur]="true"
         (ngModelChange)="myCallback1($event)"
+        (customSelected)="customCallback($event)"
         placeholder="enter text">
         <input id="model1" [ngModel]="model1" autofocus />
       </div>
@@ -57,7 +60,7 @@ let templateStr: string = `
       
   <fieldset><legend><h2>Source as HTTP URI String</h2></legend>
     <ngui-utils-4>
-      <input  ngui-auto-complete
+      <input ngui-auto-complete
         id="model4"
         [(ngModel)]="model4"
         placeholder="Enter Address(min. 2 chars)"
@@ -95,15 +98,16 @@ let templateStr: string = `
     
   <fieldset><legend><h2>With Material Design</h2></legend>
     <ngui-utils-6>
-      <md-input ngui-auto-complete 
-        id="model6"
-        [(ngModel)]="myModel"
-        [source]="arrayOfNumbers"
-        [list-formatter]="rightAligned"
-        placeholder="amount" align="end">
-        <span md-prefix>$&nbsp;</span>
-        <span md-suffix>.00</span>
-      </md-input>
+      <md-input-container>
+        <span mdPrefix>$&nbsp;</span>
+        <input mdInput ngui-auto-complete style="border: 1px solid #ccc"
+          id="model6"
+          [(ngModel)]="myModel"
+          [source]="arrayOfNumbers"
+          [list-formatter]="rightAligned"
+          placeholder="amount" align="end"/>
+          <span mdSuffix>.00</span>
+      </md-input-container>
     </ngui-utils-6>
     <pre>{{templateStr | htmlCode:'ngui-utils-6'}}</pre>
     <pre>arrayOfNumbers: {{json(arrayOfNumbers)}}</pre>
@@ -182,14 +186,18 @@ export class DirectiveTestComponent {
     private _sanitizer: DomSanitizer ) {
   }
 
-  myCallback1(newVal) {
-    console.log("value is changed to ", newVal);
-    this.model1 = newVal;
+  customCallback(text) {
+    console.log("keyword ", text)
   }
 
-  myCallback7(newVal) {
-    console.log("value is changed to ", newVal);
-    this.model7 = newVal;
+  myCallback1(newVal1) {
+    console.log("value is changed to ", newVal1);
+    this.model1 = newVal1;
+  }
+
+  myCallback7(newVal7) {
+    console.log("value is changed to ", newVal7);
+    this.model7 = newVal7;
   }
 
   renderHero = (data: any) : SafeHtml => {